Change the location of the sarr print and add functions, also change the camera and light flags C and L to uppercase
