+++ old.rt
@@ -1,5 +1,8 @@
 R 1280 720
 
+A 0.2 255,255,255
+C 0,0,20 0,0,-1 50
+L 0,0,3 1 255,255,255
 pl 0,0,1  0,0,1 255,0,255
 pl -10,0,0 1,0,0 255,0,255
 pl 10,0,0 -1,0,0 255,0,255
