+++ src/parser/parser_light.c
@@ -31,7 +31,4 @@ int	parser_light(Scene *scene, char **data)
 	scene->lights->point =  point;
 	scene->lights->ratio =  ft_limit(0, 1, ft_ratio(data[2]));	
 	scene->n_lights++;
-	// printf("----> |%f| |%f| |%f| |%f|\n", scene->lights->point.x, scene->lights->point.y, scene->lights->point.z, scene->lights->ratio);
-	// printf("----> |%f| |%f| |%f|\n", scene->lights->color->r, scene->lights->color->g, scene->lights->color->b);
-	//exit (1);
 }
