+++ src/parser/parser_obj.c
@@ -12,20 +12,6 @@
 
 #include "../main.h"
 
-int	ft_sarrprint(char **arr)
-{
-	int	i;
-
-	if (!arr || !arr[0])
-		return (0);
-	i = -1;
-	while (arr[++i])
-	{
-		ft_putstr_fd(arr[i], 1);
-		ft_putstr_fd("\n", 1);
-	}
-	return (i);
-}
 
 int	idstr(char *alphabet[], char *str)
 {
@@ -36,35 +22,6 @@ int	idstr(char *alphabet[], char *str)
 		if (ft_strncmp(alphabet[i], str, ft_strlen(str)) == 0)
 			return (i);
 	return (9); //devolver el ultimo + 1 idx del alphabet
-}
-
-char	**ft_sarradd(char **arr, char *string)
-{
-	int		size;
-	char	**new_arr;
-
-	if (!string)
-		return (arr);
-	size = ft_sarrsize(arr) + 2;
-	new_arr = malloc(sizeof(char *) * (size));
-	if (!new_arr)
-		return (NULL);
-	size = 0;
-	if (arr)
-	{
-		while (arr[size])
-		{
-			new_arr[size] = ft_strdup(arr[size]);
-			if (!new_arr[size++])
-				return (ft_free_p2(new_arr), NULL);
-		}
-		ft_free_p2(arr);
-	}
-	new_arr[size++] = ft_strdup(string);
-	if (!new_arr[size - 1])
-		return (ft_free_p2(new_arr), NULL);
-	new_arr[size] = NULL;
-	return (new_arr);
 }
 
 /**
@@ -93,8 +50,8 @@ char **init_args()
 	char **alphabet;
 	alphabet = ft_sarradd(alphabet, "R"); //posicion 0 Resolution
 	alphabet = ft_sarradd(alphabet, "A");//posicion 1 Ambient
-	alphabet = ft_sarradd(alphabet, "c");//posicion 2	camara
-	alphabet = ft_sarradd(alphabet, "l");//posicion 2	Light
+	alphabet = ft_sarradd(alphabet, "C");//posicion 2	camara
+	alphabet = ft_sarradd(alphabet, "L");//posicion 2	Light
 	alphabet = ft_sarradd(alphabet, "pl");//posicion 4 Plane
 	alphabet = ft_sarradd(alphabet, "sp"); //posicion 5 Sphere
 	alphabet = ft_sarradd(alphabet, "sq");//posicion 6 Square
